made messages show up
