+++ src/components/messages/MessageCard.js
@@ -2,9 +2,9 @@ import React from 'react';
 
 export const MessageCard = ({cardMessage}) => {
   return (
-    <div class='messageCard'>
-      <p class='message'>{cardMessage.message}</p>
-      <div class='timestamp'>{cardMessage.timestamp}</div>
+    <div className='messageCard'>
+      <p className='message'>{cardMessage.message}</p>
+      <div className='timestamp'>{cardMessage.timestamp}</div>
     </div>
   );
 }
