+++ src/components/messages/Messages.css
@@ -1,0 +1,67 @@
+fieldset {
+  max-width: 40em;
+  padding: 4px;
+  margin: 2em auto;
+  border: 0 none;
+  box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
+  border-radius: 5px;
+  background: aliceblue;
+}
+
+fieldset input, fieldset button, fieldset select{
+  box-sizing: border-box;
+  padding: .2em .4em;
+  margin: .2em 0;
+  outline: none;
+  box-shadow: none;
+}
+
+fieldset button {
+  max-width: 9em;
+  padding: .5em;
+  margin: .3em auto;
+  background-color: cornflowerblue;
+  border: none;
+  color: cornsilk;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 14px;
+  border-radius: 2px;
+  width: 100%;
+  cursor: pointer;
+}
+
+fieldset label {
+  box-sizing: border-box;
+  user-select: none;
+  cursor: pointer;
+  text-align: right;
+}
+fieldset input:focus + label {
+  color: #933;
+}
+
+fieldset .formgrid {
+  display: grid;
+  grid-template-columns: 1fr 1em 2fr;
+  grid-gap: .3em .6em;
+  grid-auto-flow: dense;
+  align-items: center;
+}
+
+fieldset input, fieldset button, fieldset select {
+  grid-column: 2 /4;
+  width: auto;
+  margin: 0
+}
+
+fieldset .alignRight {
+  text-align: right;
+}
+
+.messagePrint {
+  margin-top: 20em;
+  margin-left: 5em;
+  margin-right: 5em;
+}
